mkfs: prevent an integer overflow when computing the FAT length
The FAT length field on disk is 4 bytes. The computation of the FAT
length, roughly "num_cluster * 4" can easily overflow those 4 bytes if
the number of cluster is high.

This patch adds a sanity check before computing the FAT length.

This bug was observed when formatting a large HDD (16 TB) with a small
cluster size (8 KiB). `mkfs.exfat` succeeded but Linux refused to
mount the file system and reported an error with the FAT
length ("bogus fat length").

Signed-off-by: Christophe Vu-Brugier <christophe.vu-brugier@seagate.com>