fsck: disable "rescue orphaned clusters" option and fix upcase table decompression
Too many files could be created in LOST+FOUND
directory, if NoFatChain files are corrupted.
Because a file could be created per cluster
allocated to the NotFatChain file.
And write bitmap for clusters marked as
free but allocated to files.

Fix upcase table decompression error, which
is not reproducible, but fix the error anyway.

Signed-off-by: Hyunchul Lee <hyc.lee@gmail.com>