lib: support get the first empty dentry set position
This commit adds ->in.dentry_count to exfat_lookup_filter.

If ->in.dentry_count is 0, the function of exfat_lookup_dentry_set()
has no change. After calling exfat_lookup_dentry_set(), if the dentry
set is not found, ->out.dev_offset will be the position of the last
empty dentry set.

If ->in.dentry_count is not 0, the function of exfat_lookup_dentry_set()
will be changed. After calling exfat_lookup_dentry_set(), if the dentry
set is not found, ->out.dev_offset will be the position of the first
empty dentry set with ->in.dentry_count empty dentries.

Signed-off-by: Yuezhang Mo <Yuezhang.Mo@sony.com>
Reviewed-by: Andy Wu <Andy.Wu@sony.com>
Reviewed-by: Aoyama Wataru <wataru.aoyama@sony.com>
Acked-by: Hyunchul Lee <hyc.lee@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>